GO-4969: Move key.go to util
