+++ core/api/util/key.go
@@ -1,4 +1,4 @@
-package object
+package util
 
 import (
 	"strings"
